Migrate listTables() to SchemaCollection
Migrate the methods + tests for listTables() to the new schema
system.
